+++ src/Plumix/Widgets/App.cs
@@ -622,8 +622,6 @@ public sealed class WidgetsApp : StatefulWidget
                 result = new CheckedModeBanner(result);
             }
 
-            result = Overlay.Wrap(result);
-
             result = new Focus(
                 canRequestFocus: false,
                 onKeyEvent: HandleRootKeyEvent,